moved some helper method to its own class
