+++ Plugin/GameInfo/IGameInfo.cs
@@ -1,0 +1,6 @@
+namespace UniTASPlugin.GameInfo;
+
+public interface IGameInfo
+{
+    public string UnityVersion { get; }
+}
